requirements/base.txt
attrdict==2.0.1
coveralls==1.8.0
django==2.2.2
django-model-utils==3.1.2
django-admin-row-actions==0.0.5
git+https://github.com/mgaitan/django-sneak.git@#egg=django-sneak
git+https://github.com/mgaitan/django-admincommand.git@filefield#egg=django-admincommand
git+https://github.com/mgaitan/django-versatileimagefield@one-dimension-resize#egg=django-versatileimagefield
django-annoying==0.10.4
django-anymail==6.0.1
django-dbbackup==3.2.0
django-fancy-cache==0.11.0
django-material==1.5.6
django-extensions==2.1.7
django-formtools==2.1
djangorestframework==3.9.4
html2text==2016.9.19
django-geojson==2.10.0
git+https://github.com/mgaitan/django-hijack.git#egg=django-hijack
django-leaflet==0.22.0
django-localflavor==2.2
IPython==7.5.0
jsonfield==2.0.2
nameparser==0.5.3
phonenumbers==8.6.0
Pillow==4.2.1
pyexcel-xlsx==0.4.1
Fabric3==1.14.post1
easyimap==0.6.3
factory_boy==2.12.0
pytest-django==3.5.0
pytest-cov==2.7.1
psycopg2_binary==2.8.2
sentry-sdk==0.7.6
django-excel==0.0.10
pyexcel==0.5.13
